requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
ecdsa==0.17.0
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.10.2
hyperframe==6.0.1
idna==2.10
passlib==1.7.4
priority==1.3.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.19
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wsproto==1.0.0
